% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{methods}
\name{addValidationData}
\alias{addValidationData}
\alias{addValidationData,ClustAllObject,numericOrCharacter-method}
\title{Add the validation data into the ClustAllObject}
\usage{
addValidationData(Object, dataValidation)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{dataValidation}{numericOrCharacter}
}
\value{
\code{\link{ClustAllObject-class}} object
}
\description{
Generic function to add validation data to the
\code{\link{ClustAllObject-class}} object
}
\details{
addValidationData
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
label <- as.numeric(as.factor(wdbc$Diagnosis))
wdbc <- wdbc[,-c(1, 2)] # delete patients IDs & label
obj_noNA <- createClustAll(data = wdbc)
obj_noNA <- addValidationData(Object = obj_noNA,
                              dataValidation = label)
}
\seealso{
\code{\link{ClustAllObject-class}}
}
