% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{characterOrNA}
\alias{characterOrNA}
\alias{characterOrNA-class}
\title{characterOrNA
Class union of character, null and missing}
\value{
characterOrNA class object
}
\description{
Contains either character, NULL or missing object
}
