% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{ClustAllObject-class}
\alias{ClustAllObject-class}
\title{ClustAllObject}
\value{
ClustAllObject class object.
}
\description{
This class contains the original data and the imputed datasets
needed to run the ClustAll pipeline. The results of all the stratifications
are stored in summary_clusters. From the stratifications that pass the
bootstrapping, a matrix with the Jaccard distances is calculated and stored
in JACCARD_DISTANCE_F.
}
\section{Slots}{

\describe{
\item{\code{data}}{Data Frame of the input data after applying one-hot encoding to the
categorical variables and extracting the validation (true label) column.}

\item{\code{dataOriginal}}{Data Frame of the input data.}

\item{\code{dataImputed}}{Mids object derived from the mice package that stores the
imputed data, in case imputation was applied. Otherwise NULL.}

\item{\code{dataValidation}}{A vector in the case there is a validation (true label)
column in the input data. This information can be added later with
\code{\link{addValidationData}}. Otherwise NULL.}

\item{\code{nImputation}}{Number of imputations performed.}

\item{\code{processed}}{A boolean. TRUE if \code{\link{runClustAll}} has been
executed. Otherwise FALSE.}

\item{\code{summary_clusters}}{List with the resulting stratifications
for each combination of clustering methods (distance + clustering algorithm)
and depth, in case \code{\link{runClustAll}} has been executed previously.
Otherwise NULL.}

\item{\code{JACCARD_DISTANCE_F}}{Matrix containing the Jaccard distances derived from
the robust stratifications after applying the bootstrapping if
\code{\link{runClustAll}} has been executed previously. Otherwise NULL.}
}}

