% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_runClustAll_internal.R
\name{cluster_similarity_adapt}
\alias{cluster_similarity_adapt}
\title{The following code has been adapted from the clusteval v0.1 package created
by John A. Ramey.
\url{https://cran.r-project.org/src/contrib/Archive/clusteval/}}
\usage{
cluster_similarity_adapt(
  labels1,
  labels2,
  similarity = c("jaccard", "rand"),
  method = "independence"
)
}
\description{
The following code has been adapted from the clusteval v0.1 package created
by John A. Ramey.
\url{https://cran.r-project.org/src/contrib/Archive/clusteval/}
}
\author{
John A. Ramey
}
