% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_createClustAll.R
\name{createClustAll}
\alias{createClustAll}
\alias{createClustAll,data.frame,numericOrNA,ANY,characterOrNA-method}
\title{createClustAll: Creates ClustAllObject}
\usage{
createClustAll(data=data,
                      nImputation=NULL,
                      dataImputed=NULL,
                      colValidation=NULL)
}
\arguments{
\item{data}{Data Frame of the using data. It may contain missing values (NA).}

\item{nImputation}{Number of imputations to be computed in case the data
contains NAs and impued data is not available.}

\item{dataImputed}{mids object created with mice package. The input data
for the imputation and the data must be the same.}

\item{colValidation}{vector with the referece labeling of the original
dataset provided in “data” (if available). Default is NULL.}
}
\value{
\code{\link{ClustAllObject-class}} object.
}
\description{
Creates the ClustAllObject object. Applies one-hot encoding to columns with
categorical values from the input data and extracts the validation column if
available. It performs data imputation if the dataset contains missing values
and no imputed dataset is provided.
In total there are 3 scenarios when we create the object:
- Scenario 1: data does not contain missing values.
-Scenario 2: data contains NAs and there is no imputed data. Then, it
performs the imputations automatically.
- Scenario 3: data contains NAs and imputed data is provided manually.
Once \code{\link{ClustAllObject-class}} has been created, the ClustALL
pipeline can be run executing \code{\link{runClustAll}}.
}
\examples{
# Scenario 1: data does not contain missing values
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- wdbc[,-c(1,2)]
obj_noNA <- createClustAll(data = wdbc)

# Scenario 2: data contains NAs and there is no imputed data.
# Then it performs the imputations automatically
data("BreastCancerWisconsinMISSING", package = "ClustAll")
\donttest{
obj_NA <- createClustAll(wdbcNA, nImputation = 2)
}
# Scenario 3: data contains NAs and imputed data is provided manually
data("BreastCancerWisconsinMISSING", package = "ClustAll")
ini <- mice::mice(wdbcNA, maxit = 0, print = FALSE)
pred <- ini$pred # predictor matrix
pred["radius1", c("perimeter1", "area1", "smoothness1")] <- 0 # example of
# how to remove predictors
\donttest{
imp <- mice::mice(wdbcNA, m=5, pred=pred, maxit=5, seed=1234, print=FALSE)
obj_imp <- createClustAll(data=wdbcNA, dataImputed = imp)
}

}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
