% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{dataImputed}
\alias{dataImputed}
\alias{dataImputed,ClustAllObject-method}
\title{Retrieve the imputed data from ClustAllObject}
\usage{
dataImputed(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
Mids class object with the imputed data or NULL if imputation was
not required
}
\description{
Generic function to retrieve the imputed data obtained in
\code{\link{createClustAll}} from a \code{\link{ClustAllObject-class}} object
}
\examples{
data("BreastCancerWisconsinMISSING", package = "ClustAll")
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=-ID)
obj_NA <- createClustAll(data = wdbcNA, colValidation = "Diagnosis",
                         dataImputed = wdbcMIDS)
dataImputed(obj_NA)
}
\seealso{
\code{\link{createClustAll}}, \code{\link{ClustAllObject-class}},
\code{\link{runClustAll}}
}
