% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{initialize,ClustAllObject-method}
\alias{initialize,ClustAllObject-method}
\title{initializeClustAllObject}
\usage{
\S4method{initialize}{ClustAllObject}(
  .Object,
  data,
  dataOriginal,
  dataImputed,
  dataValidation,
  nImputation,
  processed,
  summary_clusters,
  JACCARD_DISTANCE_F
)
}
\arguments{
\item{.Object}{initializing object}

\item{data}{Data Frame of the input data after applying one-hot encoding to the
categorical variables and extracting the validation (true label) column.}

\item{dataOriginal}{Data Frame of the input data.}

\item{dataImputed}{Mids object derived from the mice package that stores the
imputed data, in case imputation was applied. Otherwise NULL.}

\item{dataValidation}{A vector in the case there is a validation (true label)
column in the input data. This information can be added later with
\code{\link{addValidationData}}. Otherwise NULL.}

\item{nImputation}{Number of imputations performed.}

\item{processed}{A boolean. TRUE if \code{\link{runClustAll}} has been
executed. Otherwise FALSE.}

\item{summary_clusters}{List with the resulting stratifications
for each combination of clustering methods (distance + clustering algorithm)
and depth, in case \code{\link{runClustAll}} has been executed previously.
Otherwise NULL.}

\item{JACCARD_DISTANCE_F}{Matrix containing the Jaccard distances derived from
the robust stratifications after applying the bootstrapping if
\code{\link{runClustAll}} has been executed previously. Otherwise NULL.}
}
\value{
ClustAllObject class object.
}
\description{
constuctor for \code{\link{ClustAllObject-class}}
}
