% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{listOrNULL}
\alias{listOrNULL}
\alias{listOrNULL-class}
\title{Class Union listOrNULL}
\value{
listOrNULL class object
}
\description{
Contains either list, NULL or missing object
}
\details{
Class union of list, null and missing
}
