% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{logicalOrNA}
\alias{logicalOrNA}
\alias{logicalOrNA-class}
\title{logicalOrNA}
\description{
Contains either logical, NULL or missing object
}
\details{
Class union of logical, null or missing
}
