% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{matrixOrNULL-class}
\alias{matrixOrNULL-class}
\title{Class Union: matrixOrNULL}
\description{
This class union allows for flexibility in method signatures and slot
definitions by accepting either a matrix or NULL. It's particularly useful
when a slot or function parameter might contain a matrix of data but could
also be empty or explicitly set to NULL.
}
\details{
The matrixOrNULL class union includes:
\itemize{
  \item matrix: A standard R matrix object
  \item NULL: Representing an empty or unset value
}

This union is useful in scenarios where:
\itemize{
  \item A function might return a matrix of results or NULL if no results are available
  \item A slot in an S4 object could contain a matrix of data or be empty
  \item A function parameter could accept a matrix input, but also work with default settings if nothing is provided
}
}
\seealso{
\code{\link{setClassUnion}}, \code{\link{ClustAllObject-class}}
}
