% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{matrixOrNULL}
\alias{matrixOrNULL}
\alias{matrixOrNULL-class}
\title{matrixOrNULL}
\value{
matrixOrNULL class object
}
\description{
Contains either matrix or NULL object
}
\details{
Class union of matrix, null and missing
}
