% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{numericOrNA}
\alias{numericOrNA}
\alias{numericOrNA-class}
\title{Class Union numericOrNA}
\value{
numericOrNA class object
}
\description{
Contains either numeric, NULL or missing object
}
\details{
Class union of numeric, null or missing
}
