% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{data}
\name{obj_noNAno1Validation}
\alias{obj_noNAno1Validation}
\title{obj_noNAno1Validation: Processed wdbc dataset for testing purposed}
\format{
A processed ClustAllObject
}
\usage{
data("testData", package = "ClustAll")
}
\value{
ClustAllObject Object
}
\description{
Processed wdbc as appear in vignette, with no validation data
}
\keyword{datasets}
