% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{show,ClustAllObject-method}
\alias{show,ClustAllObject-method}
\title{show method for ClustAllObject}
\usage{
\S4method{show}{ClustAllObject}(object)
}
\arguments{
\item{object}{\code{\link{ClustAllObject-class}} object}
}
\value{
summarize information about the object
}
\description{
Show method for a \code{\link{ClustAllObject-class}} object
}
