% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showDataImputed}
\alias{showDataImputed}
\alias{showDataImputed,ClustAllObject-method}
\title{Retrieve the imputed data from ClustAllObject}
\usage{
showDataImputed(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
Mids class object with the imputed data or NULL value if no imputations have been made
}
\description{
Generic function to retrieve the imputed data obtained in \code{\link{createClustAll}} from a \code{\link{ClustAllObject-class}} object
}
\examples{
# MISSING EXAMPLES
}
\seealso{
\code{\link{createClustAll}}, \code{\link{ClustAllObject-class}}, \code{\link{runClustAll}}
}
