% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showNumberImputations}
\alias{showNumberImputations}
\alias{showNumberImputations,ClustAllObject-method}
\title{Retrieve the number of imputations applied at the imputation step from
ClustAllObject}
\usage{
showNumberImputations(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
Numeric vector that contains the number of imputations. 0 in the
case of no imputations were required
}
\description{
Generic function to retrieve the number of imputations in
\code{\link{createClustAll}} from a \code{\link{ClustAllObject-class}} object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc$Diagnosis <- as.numeric(as.factor(wdbc$Diagnosis))
wdbc <- wdbc[,-c(1)] # delete patients IDs
obj_noNA <- createClustAll(data = wdbc, colValidation = "Diagnosis")
showNumberImputations(obj_noNA)
}
\seealso{
\code{\link{createClustAll}}, \code{\link{ClustAllObject-class}},
\code{\link{runClustAll}}
}
