% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showSummaryClusters}
\alias{showSummaryClusters}
\alias{showSummaryClusters,ClustAllObject-method}
\title{Retrieve the list with matrieces for different cluster methods from ClustAllObject}
\usage{
showSummaryClusters(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
List of matrices with all the clustering methods or NULL value if runClustAll method have not been runned
}
\description{
Generic function to retrieve the list with matrieces for different cluster methods results of \code{\link{runClustAll}} from a \code{\link{ClustAllObject-class}} object
}
\examples{
# MISSING EXAMPLES
}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
