% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showValidationData}
\alias{showValidationData}
\alias{showValidationData,ClustAllObject-method}
\title{Retrieve the original data labelling from ClustAllObject}
\usage{
showValidationData(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
numeric vector if true labels have been added. Otherwise NULL
}
\description{
Generic function to retrieve numeric vector if it has been added with the
true labels from a \code{\link{ClustAllObject-class}} object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=-ID)
obj_noNA <- createClustAll(data = wdbc, colValidation="Diagnosis")
showValidationData(obj_noNA)
}
\seealso{
\code{\link{ClustAllObject-class}}
}
