% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{data}
\name{wdbc}
\alias{wdbc}
\title{wdbc: Diagnostic Wisconsin Breast Cancer Database.}
\format{
A data frame with 660 rows and 31 variables
}
\source{
<https://archive.ics.uci.edu/dataset/17/breast+cancer+wisconsin+diagnostic>

\itemize{
  \item Diagnosis Label says tumor is malingnant or benignant
  \item radius. Mean of distances from the center to points on the perimeter
  \item perimeter
  \item area
  \item smoothness. Local variation in radius lengths
  \item compactness. (Perimeter^2 / Area) - 1.0
  \item concavity. Severity of concave portions of the contour
  \item concave points. Number of concave portions of the contour
  \item symmetry.
  \item fractal dimension. “Coastline approximation” - 1.
}
}
\usage{
data("BreastCancerWisconsin", package = "ClustAll")
}
\description{
A dataset containing Features are computed from a digitized image of a fine
needle aspirate (FNA) of a breast mass.
They describe characteristics of the cell nuclei present in the image.
}
\details{
The dataset comprises two types of features —categorical and numerical—
derived from a digitized image of a fine needle aspirate (FNA) of a breast
mass from 659 patients. Each patient is characterized by 31 features (10x3)
and belongs to one of two target classes: ‘malignant’ or ‘benign’.
}
\keyword{datasets}
