% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{data}
\name{wdbcMIDS}
\alias{wdbcMIDS}
\title{wdbcMIDS: Diagnostic Wisconsin Breast Cancer Database with imputed values}
\format{
A data frame with 660 rows and 31 variables
}
\usage{
data("BreastCancerWisconsinMISSING", package = "ClustAll")
}
\value{
wdbcMIDS dataset
}
\description{
The "wdbcMIDS" dataset is an imputed version of the "wdbcNA" dataset, which is
a modified version of the Breast Cancer Wisconsin (Diagnostic) dataset
(\code{\link{wdbc}}) with missing values introduced at random. The missing values
have been imputed using the Multiple Imputation by Chained Equations (MICE)
algorithm, resulting in a "mids" object manually.
}
\details{
For more information on the original dataset and the version with missing values,
please refer to the documentation of the "wdbc" (\code{\link{wdbc}}).
}
\keyword{datasets}
