% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{data}
\name{wdbcNA}
\alias{wdbcNA}
\title{wdbcNA: Diagnostic Wisconsin Breast Cancer Database with missing values}
\format{
A data frame with 660 rows and 31 variables
}
\usage{
data("BreastCancerWisconsinMISSING", package = "ClustAll")
}
\description{
We introduced random missing values to the wdbc dataset. \code{\link{wdbc}}
}
\keyword{datasets}
