% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{addValidationData}
\alias{addValidationData}
\alias{addValidationData,ClustAllObject-method}
\title{Add the validation data into the ClustAllObject}
\usage{
addValidationData(Object, dataValidation)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{dataValidation}{numeric vector with the validation data}
}
\value{
\code{\link{ClustAllObject-class}} object
}
\description{
Generic function to add validation data to the \code{\link{ClustAllObject-class}} object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc$Diagnosis <- as.numeric(as.factor(wdbc$Diagnosis))
wdbc <- wdbc[,-c(1)] # delete patients IDs
obj_noNA <- createClustAll(data = wdbc, colValidation = "Diagnosis")
obj_noNA1 <- addValidationData(Object = obj_noNA, dataValidation = wdbc$Diagnosis)
}
\seealso{
\code{\link{ClustAllObject-class}}
}
