% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{addValidationData}
\alias{addValidationData}
\alias{addValidationData,ClustAllObject-method}
\title{Add the validation data into the ClustAllObject}
\usage{
addValidationData(Object, dataValidation)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{dataValidation}{numeric vector with the validation data}
}
\value{
\code{\link{ClustAllObject-class}} object
}
\description{
Generic function to add validation data to the \code{\link{ClustAllObject-class}} object
}
\examples{
# MISSING EXAMPLES
}
\seealso{
\code{\link{ClustAllObject-class}}
}
