% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{methods}
\name{addValidationData}
\alias{addValidationData}
\alias{addValidationData,ClustAllObject,numericOrCharacter-method}
\title{addValidationData}
\usage{
addValidationData(Object, dataValidation)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object.}

\item{dataValidation}{A numeric or character vector with the
validation data (true labels).
The length of the vector must be the same as input data used in
\code{\link{createClustAll}}.}
}
\value{
\code{\link{ClustAllObject-class}} object.
}
\description{
Generic function to add validation data to the
\code{\link{ClustAllObject-class}} object.
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
label <- as.numeric(as.factor(wdbc$Diagnosis))
wdbc <- wdbc[,-c(1, 2)] # delete patients IDs & label
obj_noNA <- createClustAll(data = wdbc)
obj_noNA <- addValidationData(Object = obj_noNA,
                              dataValidation = label)
}
\seealso{
\code{\link{ClustAllObject-class}}
}
