% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{ClustAllObject-class}
\alias{ClustAllObject-class}
\title{ClustAllObject}
\description{
Stores the original data used, the imputed datasets and the
results of the ClustAll pipeline.
}
\section{Slots}{

\describe{
\item{\code{data}}{Data Frame of the data used. Maybe modified from the input
data.}

\item{\code{dataOriginal}}{Data Frame of the original data introduced.}

\item{\code{dataImputed}}{midsOrNULL. Mids object derived from the
mice package that stores the imputed data, in case
imputation was applied. Otherwise NULL.}

\item{\code{dataValidation}}{labelling numericOrNA. Original data labelling.}

\item{\code{nImputation}}{Number of multiple imputations to be applied.}

\item{\code{processed}}{Logical if the ClustAll pipeline has been executed previously}

\item{\code{summary_clusters}}{listOrNULL. List with the resulting stratifications
for each combination of clustering methods (distance + clustering algorithm)
and depth, in case ClustAll pipeline has been executed previously.
Otherwise NULL.}

\item{\code{JACCARD_DISTANCE_F}}{matrixOrNULL. Matrix containing the Jaccard
distances derived from the robust populations stratifications if ClustAll
pipeline has been executed previously. Otherwise NULL.}
}}

