% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{ClustAllObject-class}
\alias{ClustAllObject-class}
\title{ClustAllObject}
\description{
Stores the using data, imputations and results of the ClustAll pipeline.
}
\section{Slots}{

\describe{
\item{\code{data}}{Data Frame of the using data}

\item{\code{dataImputed}}{midsOrNULL. Mids object from mice package that stores imputed data in case imputations have been made. Otherwise NULL value}

\item{\code{dataValidation}}{numericOrNA. Numeric vector of the TRUE labels}

\item{\code{nImputation}}{Numeric giving the number of imputations}

\item{\code{processed}}{Logical value if the ClustAll pipeline have been runned}

\item{\code{summary_clusters}}{listOrNULL. List of matrices of the clustering methods in case ClustAll pipeline have been runned. Othrwise NULL value}

\item{\code{JACCARD_DISTANCE_F}}{matrixOrNULL. Matrix containing the robust values of Jaccard distances if the ClustAll pipeline have been runned. Otherwise NULL value}
}}

