% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{cluster2data}
\alias{cluster2data}
\alias{cluster2data,ClustAllObject,character-method}
\title{cluster2data}
\usage{
cluster2data(Object,
                    stratificationName)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{stratificationName}{Character vector with one or more stratification names}
}
\value{
data.frame
}
\description{
Returns the original data in a dataframe, including the selected robust
stratification(s) as varaibles. The representative stratification names can
be obtained using the method. \code{\link{resStratification}}
}
\examples{
\dontrun{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
resStratification(Object = obj_noNA1, population = 0.05,
                  stratification_similarity = 0.88, all = FALSE)
df <- cluster2data(Object = obj_noNA1,
                   stratificationName = c("cuts_a_1","cuts_b_5","cuts_a_5"))
}
}
\seealso{
\code{\link{resStratification}},\code{\link{plotJACCARD}},
\code{\link{ClustAllObject-class}}
}
