% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{cluster2data}
\alias{cluster2data}
\alias{cluster2data,ClustAllObject,character-method}
\title{cluster2data: Export selected stratification(s)}
\usage{
cluster2data(Object,
                    stratificationName)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object.}

\item{stratificationName}{Name of the stratification(s) to be exported.}
}
\value{
Returns the original Data Frame with additional column(s)
corresponding to the selected stratification(s).
}
\description{
Returns the original input data in a Data Frame, appending the selected
robust stratification(s) as additional columns. The names of the representative
stratifications can be obtained using the \code{\link{resStratification}}
method.
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
\donttest{
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
resStratification(Object = obj_noNA1, population = 0.05,
                  stratification_similarity = 0.88, all = FALSE)
df <- cluster2data(Object = obj_noNA1,
                   stratificationName = c("cuts_a_1","cuts_b_5","cuts_a_5"))
}
}
\seealso{
\code{\link{resStratification}},\code{\link{plotJACCARD}},
\code{\link{ClustAllObject-class}}
}
