% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{cluster2data}
\alias{cluster2data}
\alias{cluster2data,ClustAllObject-method}
\title{cluster2data}
\usage{
cluster2data(Object,
       clusterName)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{clusterName}{Character vector with one or more stratification names}
}
\value{
data.frame
}
\description{
Returns the data frame of the original data using which the clustering of the selected stratification(s) are included as varibles. The representative stratification names can be obtained using the method \code{\link{resStratification}}
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- wdbc[,-c(1,2)]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 8)
resStratification(Object = obj_noNA1, population = 0.05, stratification_similarity = 0.88, all = FALSE)
df <- cluster2data(Object = obj_noNA1, clusterName = c("cuts_c_3","cuts_a_9","cuts_b_13"))
}
\seealso{
\code{\link{resStratification}},\code{\link{plotJACCARD}},\code{\link{ClustAllObject-class}}
}
