% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{extractData}
\alias{extractData}
\alias{extractData,ClustAllObject-method}
\title{extractData}
\usage{
extractData(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object.}
}
\value{
List with the information of original, modified and imputed data.
}
\description{
Generic function to retrieve all the data used in
\code{\link{createClustAll}} from a \code{\link{ClustAllObject-class}}
object.
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=-ID)
\donttest{
obj_noNA <- createClustAll(data = wdbc, colValidation = "Diagnosis")
extractData(obj_noNA)
}
}
\seealso{
\code{\link{createClustAll}}, \code{\link{ClustAllObject-class}},
\code{\link{runClustAll}}
}
