% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{extractResults}
\alias{extractResults}
\alias{extractResults,ClustAllObject-method}
\title{extractResults}
\usage{
extractResults(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object.}
}
\value{
List with all the names of generated stratifications and the
statically robust ones.
}
\description{
Generic function to retrieve all the results after processing
\code{\link{runClustAll}} from a \code{\link{ClustAllObject-class}} object.
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=-ID)
\donttest{
obj_noNA <- createClustAll(data = wdbc, colValidation = "Diagnosis")
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
extractResults(obj_noNA1)
}

}
\seealso{
\code{\link{createClustAll}}, \code{\link{ClustAllObject-class}},
\code{\link{runClustAll}}
}
