% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{initialize,ClustAllObject-method}
\alias{initialize,ClustAllObject-method}
\title{initializeClustAllObject}
\usage{
\S4method{initialize}{ClustAllObject}(
  .Object,
  data,
  dataImputed,
  dataValidation,
  nImputation,
  processed,
  summary_clusters,
  JACCARD_DISTANCE_F
)
}
\description{
constuctor for \code{\link{ClustAllObject-class}}
}
