% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{isProcessed}
\alias{isProcessed}
\alias{isProcessed,ClustAllObject-method}
\title{Retrieve logical value if runClustAll have been runned from ClustAllObject}
\usage{
isProcessed(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
TRUE if runClustAll have been runned. Otherwise FALSE
}
\description{
Generic function to retrieve the logical value if \code{\link{runClustAll}} have been runned from a \code{\link{ClustAllObject-class}} object
}
\examples{
# MISSING EXAMPLES
}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
