% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{isProcessed}
\alias{isProcessed}
\alias{isProcessed,ClustAllObject-method}
\title{Retrieve logical if runClustAll has been executed considering
ClustAllObject as input}
\usage{
isProcessed(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
TRUE if runClustAll has been already executed. Otherwise FALSE
}
\description{
Generic function to retrieve the logical if \code{\link{runClustAll}} have
been runned from a \code{\link{ClustAllObject-class}} object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
isProcessed(obj_noNA)
}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
