% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{JACCARD_DISTANCE_F}
\alias{JACCARD_DISTANCE_F}
\alias{JACCARD_DISTANCE_F,ClustAllObject-method}
\title{Retrieve the matrix with the Jaccard distances derived from the robust
populations stratifications in ClustAllObject}
\usage{
JACCARD_DISTANCE_F(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
Matrix containing the Jaccard distances derived from the robust
populations stratifications or NULL if runClustAll method has not been
executed yet
}
\description{
Generic function to retrieve the matrix with the Jaccard distances derived
from the robust populations stratifications in\code{\link{runClustAll}} from
a \code{\link{ClustAllObject-class}} object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = FALSE)
JACCARD_DISTANCE_F(obj_noNA1)
}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
