% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{class}
\name{numericOrCharacter}
\alias{numericOrCharacter}
\alias{numericOrCharacter-class}
\title{numericOrCharacter}
\value{
numericOrCharacter class object
}
\description{
Contains either numeric or character object
}
\details{
Class union of numeric and character
}
