% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{data}
\name{obj_noNA1}
\alias{obj_noNA1}
\title{obj_noNA1: Processed wdbc dataset for testing purposed}
\format{
A processed ClustAllObject
}
\usage{
data("testData", package = "ClustAll")
}
\value{
ClustAllObject Object
}
\description{
Processed wdbc as appear in vignette
}
\keyword{datasets}
