% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{plotJACCARD}
\alias{plotJACCARD}
\alias{plotJACCARD,ClustAllObject,logicalOrNA,numericOrNA-method}
\title{plotJACCARD: Heatmap of robust stratification distances based on
 Jaccard similarity}
\usage{
plotJACCARD(Object,
                   paint=TRUE,
                   stratification_similarity=0.7)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object.}

\item{paint}{TRUE for painting a square Logical vector with the annotation for the different
stratifications.}

\item{stratification_similarity}{Numeric value representing the minimum
Jaccard distance required to consider two stratifications as similar.
The default is 0.7.}
}
\value{
Plot of a heatmap of the correlation matrix displaying Jaccard
distances between statistically robust stratifications.
}
\description{
This function plots the correlation matrix heatmap showing the Jaccard
distances between robust stratifications.
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
\donttest{
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
plotJACCARD(obj_noNA1, paint = TRUE, stratification_similarity = 0.9)
}
}
\seealso{
\code{\link{resStratification}},\code{\link{cluster2data}},
\code{\link{ClustAllObject-class}}
}
