% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{plotJACCARD}
\alias{plotJACCARD}
\alias{plotJACCARD,ClustAllObject,logicalOrNA,numericOrNA-method}
\title{Correlation matrix heatmap showing the Jaccard distance between
robust stratifications in the ClustAllObject}
\usage{
plotJACCARD(Object,
                   paint=TRUE,
                   stratification_similarity=0.7)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{paint}{Logical vector with the annotation for the different
stratifications}

\item{stratification_similarity}{The minimum Jaccard Distance value to
consider two stratifications similar. Default is 0.7.}
}
\value{
plot
}
\description{
This function plots the correlation matrix heatmap showing the Jaccard
Distance between robust stratifications
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
plotJACCARD(obj_noNA1, paint = TRUE, stratification_similarity = 0.9)
}
\seealso{
\code{\link{resStratification}},\code{\link{cluster2data}},
\code{\link{ClustAllObject-class}}
}
