% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{plotJACCARD}
\alias{plotJACCARD}
\alias{plotJACCARD,ClustAllObject-method}
\title{Plots correlation Jaccard Distances from ClustAllObject}
\usage{
plotJACCARD(Object,
       paint=TRUE,
       cluster_similarity=0.7)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{paint}{Logical vector to paint the different group of clusters in the plot}

\item{cluster_similarity}{the minimum values to consider two groups similar in Jaccard distances. As default 0.7.}
}
\value{
plot
}
\description{
This function plots the correlation Jaccard Distances
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- wdbc[,-c(1,2)]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 8)
plotJACCARD(obj_noNA1, paint = TRUE, cluster_similarity = 0.88)

}
\seealso{
\code{\link{resStratification}},\code{\link{cluster2data}},\code{\link{ClustAllObject-class}}
}
