% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{plotSANKEY}
\alias{plotSANKEY}
\alias{plotSANKEY,ClustAllObject-method}
\title{Plots Sankey diagram from two selected groups from ClustAllObject}
\usage{
plotSANKEY(Object,
                  clusters)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{clusters}{Character vector with the names of two clusters. Check resStratification to obtain stratification names.}

\item{validationData}{Logical value to use validation data to compare with the selected stratification}
}
\value{
plot
}
\description{
This function plots the Sankey diagram of two selected clusters
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- wdbc[,-c(1,2)]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 8)
resStratification(Object = obj_noNA1, population = 0.05, stratification_similarity = 0.88, all = F)
plotSANKEY(Object = obj_noNA1, clusters = c("cuts_c_3","cuts_a_9"))
plotSANKEY(Object = obj_noNA1, clusters = c("cuts_c_3","cuts_b_13"))

}
\seealso{
\code{\link{resStratification}},\code{\link{cluster2data}},\code{\link{ClustAllObject-class}}
}
