% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{processed}
\alias{processed}
\alias{processed,ClustAllObject-method}
\title{processed accession method}
\usage{
processed(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object.}
}
\value{
TRUE if runClustAll has been executed. Otherwise FALSE.
}
\description{
Generic function to know if \code{\link{ClustAllObject-class}} has been
processed. TRUE if \code{\link{runClustAll}} has been executed. Otherwise
FALSE.
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
processed(obj_noNA)
}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
