% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{resStratification}
\alias{resStratification}
\alias{resStratification,ClustAllObject,numericOrNA,logicalOrNA,numericOrNA-method}
\title{resStratification: Show the representative stratifications}
\usage{
resStratification(Object,
                         population=0.05,
                         all=FALSE,
                         stratification_similarity=0.7)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object.}

\item{population}{Numeric vector specifying the minimum percentage of the
total population that a stratification must have to be considered valid.}

\item{all}{Logical vector indicating whether to return all representative
stratifications for each group of clusters. If FALSE, only the centroid
(representative) stratification of each group is returned.}

\item{stratification_similarity}{Numeric value representing the minimum
Jaccard distance required to consider two stratifications as similar.
The default is 0.7.}
}
\value{
List containing all statistically robust stratifications,
grouped by clusters formed based on the previously defined minimum cluster
similarity and population criteria. If all is set to TRUE, the function
returns all robust stratifications for each cluster. If all is set to FALSE,
only the centroid (representative) stratification for each cluster group is
returned.
}
\description{
This function returns a list of representative stratifications.
Stratifications that do not meet the minimum population threshold for each
group are discarded. The minimum population percentage can be set with a
default value of 0.05 (5%). When 'all' is TRUE, the function returns all
robust stratifications; otherwise, it returns the representative for each
group of stratification.
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
\donttest{
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
resStratification(Object = obj_noNA1, population = 0.05,
                  stratification_similarity = 0.88, all = FALSE)
}

}
\seealso{
\code{\link{plotJACCARD}},\code{\link{cluster2data}},
\code{\link{ClustAllObject-class}}
}
