% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{resStratification}
\alias{resStratification}
\alias{resStratification,ClustAllObject,numericOrNA,logicalOrNA,numericOrNA-method}
\title{Show the stratification representatives from the ClustAllObject}
\usage{
resStratification(Object,
                         population=0.05,
                         all=FALSE,
                         stratification_similarity=0.7)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{population}{Numeric vector with the minimum percentage of the total
population that a stratification must have to be considered as representative}

\item{all}{Logical vector to return all the representative stratifications
per group of clusters. If it is FALSE, only the centroid stratification of
each group of clusters is returned}

\item{stratification_similarity}{The minimum Jaccard distance value to
consider two groups similar. Default is 0.7}
}
\value{
list
}
\description{
This function returns the stratifications representatives by keeping those
clusters with a minimum percentage of the population. Default is 0.05.
It returns all the robust stratification (TRUE) or the representative for
each group of stratifications (FALSE). Default is FALSE
}
\examples{
\dontrun{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
resStratification(Object = obj_noNA1, population = 0.05,
                  stratification_similarity = 0.88, all = FALSE)
}

}
\seealso{
\code{\link{plotJACCARD}},\code{\link{cluster2data}},
\code{\link{ClustAllObject-class}}
}
