% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_runClustAll.R
\name{runClustAll}
\alias{runClustAll}
\alias{runClustAll,ClustAllObject,numericOrNA,logicalOrNA-method}
\title{ClustAll: Data driven strategy to find hidden subgroups of patients within
complex diseases using clinical data}
\usage{
runClustAll(Object,
                   threads=1,
                   simplify=FALSE)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{threads}{Numeric vector that indicates the number of cores to use}

\item{simplify}{if TRUE computes one out of four depths of the dendrogram}
}
\value{
An object of class \code{\link{ClustAllObject-class}}
}
\description{
This method runs the ClustAll pipeline
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
}
\seealso{
\code{\link{resStratification}},\code{\link{plotJACCARD}},
\code{\link{cluster2data}},\code{\link{ClustAllObject-class}}
}
