% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_runClustAll.R
\name{runClustAll}
\alias{runClustAll}
\alias{runClustAll,ClustAllObject-method,numericOrNA}
\title{ClustAll: pipeline a data driven strategy in order to find hidden subgroups.}
\usage{
runClustAll(Object,
       threads=1)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{threads}{Numeric vector giving the number of cores to use}
}
\value{
An object of class \code{\link{ClustAllObject-class}}
}
\description{
This method runs the ClustALL algorithm
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- wdbc[,-c(1,2)]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 8)

}
\seealso{
\code{\link{resStratification}},\code{\link{plotJACCARD}},\code{\link{cluster2data}},\code{\link{ClustAllObject-class}}
}
