% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showData}
\alias{showData}
\alias{showData,ClustAllObject-method}
\title{Retrieve the initial data from ClustAllObject}
\usage{
showData(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
The Data Frame with the initial data
}
\description{
Generic function to retrieve the initial data used for \code{\link{createClustAll}} from a \code{\link{ClustAllObject-class}} object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc$Diagnosis <- as.numeric(as.factor(wdbc$Diagnosis))
wdbc <- wdbc[,-c(1)] # delete patients IDs
obj_noNA <- createClustAll(data = wdbc, colValidation = "Diagnosis")
showData(obj_noNA)
}
\seealso{
\code{\link{createClustAll}}, \code{\link{ClustAllObject-class}}, \code{\link{runClustAll}}
}
