% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showDataImputed}
\alias{showDataImputed}
\alias{showDataImputed,ClustAllObject-method}
\title{Retrieve the imputed data from ClustAllObject}
\usage{
showDataImputed(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
Mids class object with the imputed data or NULL if no imputations have been made
}
\description{
Generic function to retrieve the imputed data obtained in \code{\link{createClustAll}} from a \code{\link{ClustAllObject-class}} object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc$Diagnosis <- as.numeric(as.factor(wdbc$Diagnosis))
wdbc <- wdbc[,-c(1)] # delete patients IDs
obj_noNA <- createClustAll(data = wdbc, colValidation = "Diagnosis")
showDataImputed(obj_noNA)
}
\seealso{
\code{\link{createClustAll}}, \code{\link{ClustAllObject-class}}, \code{\link{runClustAll}}
}
