% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showJaccardDistances}
\alias{showJaccardDistances}
\alias{showJaccardDistances,ClustAllObject-method}
\title{Retrieve the matrix with Jaccard distances from ClustAllObject}
\usage{
showJaccardDistances(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
Matrix with Jaccard distances or NULL value if runClustAll method have not been runned
}
\description{
Generic function to retrieve the matrix with Jaccard distances results of \code{\link{runClustAll}} from a \code{\link{ClustAllObject-class}} object
}
\examples{
# MISSING EXAMPLES
}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
