% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showNumberImputations}
\alias{showNumberImputations}
\alias{showNumberImputations,ClustAllObject-method}
\title{Retrieve the number of imputations made for the imputation step from ClustAllObject}
\usage{
showNumberImputations(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
Numeric vector giving the number of imputations. 0 in the case of no imputations
}
\description{
Generic function to retrieve the number of imputations in \code{\link{createClustAll}} from a \code{\link{ClustAllObject-class}} object
}
\examples{
# MISSING EXAMPLES
}
\seealso{
\code{\link{createClustAll}}, \code{\link{ClustAllObject-class}}, \code{\link{runClustAll}}
}
