% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showSummaryClusters}
\alias{showSummaryClusters}
\alias{showSummaryClusters,ClustAllObject-method}
\title{Retrieve the resulting stratifications for each combination of
clusterings method (distance + clustering algorithm) and depth from
ClustAllObject}
\usage{
showSummaryClusters(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
List with the resulting stratifications for each combination of
clusterings method (distance + clustering algorithm) and depth methods or
NULL if runClustAll method has not been executed yet.
}
\description{
Generic function to retrieve the resulting stratifications for each
combination of clusterings method (distance + clustering algorithm) and
depth of \code{\link{runClustAll}} from a \code{\link{ClustAllObject-class}}
object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc$Diagnosis <- as.numeric(as.factor(wdbc$Diagnosis))
wdbc <- wdbc[,-c(1)] # delete patients IDs
obj_noNA <- createClustAll(data = wdbc, colValidation = "Diagnosis")
showSummaryClusters(obj_noNA)
}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
