% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{showValidationData}
\alias{showValidationData}
\alias{showValidationData,ClustAllObject-method}
\title{Retrieve the numeric vector of validation data from ClustAllObject}
\usage{
showValidationData(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
numeric vector if true labels have been added. Otherwise NULL
}
\description{
Generic function to retrieve numeric vector if it has been added with the true labels from a \code{\link{ClustAllObject-class}} object
}
\examples{
# MISSING EXAMPLES
}
\seealso{
\code{\link{ClustAllObject-class}}
}
