% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\name{summary_clusters}
\alias{summary_clusters}
\alias{summary_clusters,ClustAllObject-method}
\title{Retrieve the resulting stratifications for each combination of
clusterings method (distance + clustering algorithm) and depth from
ClustAllObject}
\usage{
summary_clusters(Object)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}
}
\value{
List with the resulting stratifications for each combination of
clusterings method (distance + clustering algorithm) and depth methods or
NULL if runClustAll method has not been executed yet.
}
\description{
Generic function to retrieve the resulting stratifications for each
combination of clusterings method (distance + clustering algorithm) and
depth of \code{\link{runClustAll}} from a \code{\link{ClustAllObject-class}}
object
}
\examples{
data("BreastCancerWisconsin", package = "ClustAll")
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = FALSE)
summary_clusters(obj_noNA1)
}
\seealso{
\code{\link{runClustAll}}, \code{\link{ClustAllObject-class}}
}
