% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Methods.R
\name{validateStratification}
\alias{validateStratification}
\alias{validateStratification,ClustAllObject,characterOrNA-method}
\title{validateStratification}
\usage{
validateStratification(Object,
                              stratificationName)
}
\arguments{
\item{Object}{\code{\link{ClustAllObject-class}} object}

\item{stratificationName}{Character vector with the name a stratification.
Check resStratification to obtain stratification names.}
}
\value{
numeric
}
\description{
Returns the sensitivity and specifity of the selected stratification the
original lebelling. The representative stratification names can be obtained
using the method \code{\link{resStratification}}
}
\examples{
\dontrun{
data("BreastCancerWisconsin", package = "ClustAll")
label <- as.numeric(as.factor(wdbc$Diagnosis))
wdbc <- subset(wdbc,select=c(-ID, -Diagnosis))
wdbc <- wdbc[1:15,1:8]
label <- label[16:30]
obj_noNA <- createClustAll(data = wdbc)
obj_noNA1 <- runClustAll(Object = obj_noNA, threads = 1, simplify = TRUE)
resStratification(Object = obj_noNA1, population = 0.05,
                  stratification_similarity = 0.88, all = FALSE)
obj_noNA1 <- addValidationData(Object = obj_noNA1,
                               dataValidation = label)
validateStratification(obj_noNA1, "cuts_a_1")
}

}
\seealso{
\code{\link{resStratification}},\code{\link{plotJACCARD}},
\code{\link{ClustAllObject-class}}
}
