% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAll_ClustAllObject_Class.R
\docType{data}
\name{wdbcMIDS}
\alias{wdbcMIDS}
\title{wdbcMIDS: Diagnostic Wisconsin Breast Cancer Database with imputed values}
\format{
A data frame with 660 rows and 31 variables
}
\usage{
data("BreastCancerWisconsinMISSING", package = "ClustAll")
}
\description{
We introduced imputed random values to the wdbcNA dataset.
Using Mice. It is a mids object. \code{\link{wdbc}}
}
\keyword{datasets}
